Adding templates for item-template and stats template
Feat: item-template and stats-template

item-template is used to display individual todo items.
stats-template is used to populate the footer. This template displays
the number of remaining incomplete items and contains a list of
hyperlinks which will be sued to perform actions. It also contains a
button which can be used to clear all of the completed items