Create Application View
Feat: Create Application View

The AppView will handle the creation of new todos and rendering of the
initial todo list. The initialize() method will be called to do the
instantiation and listen to the events.

The el property stores the target DOM element ( section id=“todoapp”)
element in html.

The addOne() method creates an instance of TodoView view, renders it,
and appends the resulting element to the Todo list

The addAll() will iterates over all of the Todos currently in the
collection and fires addOne() for each item