Update Application View
Chore: Update Application View with new features.

1. Denishing an events hash containing declarative callbacks for DOM
events.

2. createOnEnter(): creates a new Todo model and persists it in
localStorage when user hits enter inside the</input> field and reset
the <input> field when complete to prepare for the next entry. The
model is populated by newAttributes(), which returns an object literal
composed of the title, order and completed state of the new item.

3. clearCompleted(): removes the items in the todo list that have been
marked as completed when the user clicks the clear-complied checkbox.

4. toggleAllComplete(): Allows a user to mark all of the items in the
todo list as completed by clicking the toggle-all checkbox.

5. filterOne() is bound with the Todos collection for a
“change:completed” event. This listens for changes to the completed
flag for nay model in the collection. The affected todo is passed to
the callback which triggers a custom “visible” event on the model.

6. filterAll() is bound for a filter event. Its responsibility is to
toggle which todo items are visible based on the filter currently
selected in the UI.

7. render(): The #main and #footer sections are displayed or hidden
depending on whether there are nay todos in the collection.