Create a Todo View
Feat: Create a Todo View

A TodoView is in charge of individual Todo records, making sure the
view updates when the todo does.

An initialize() constructor is set up to monitor a todo model’s change
event. When the todo gets updated, the application will re-render the
view and visually reflective its changes.

The render() method will render #item-template.

edit(): changes the current view into editing mode when a user
double-clicks on an existing item in the todo list. This allows them to
change the existing value of the item’s title attribute.

updateOnEnter(): check that the set has hit the return/enter key and
executes the close() function.

close(): trim the value of the current text in the <input> field an
save the changes to the current todo model and close editing mode.