Completing and deleting todos
Feat: Completing and deleting todos

The togglecompleted() function is invoked which calls toggle() on the
todo model.

toggle() toggles the completed status of the todo and calls save() on
the model.
The save generates a change event on the model which is bound to our
TodoView’s  render() method.
The save also results in a change:completed event on the model which is
handled by the AppView’s filterOne() method.

The clear() method is invoked which calls destroy() on the todo model.
The todo is deleted from local storage and a destroy event is triggered.